+++ pkg/apiserver/http.go
@@ -10,9 +10,15 @@ import (
 	"time"
 
 	"github.com/gin-gonic/gin"
+	sloggin "github.com/samber/slog-gin"
+	swaggerfiles "github.com/swaggo/files"
+	ginswagger "github.com/swaggo/gin-swagger"
+	"go.opentelemetry.io/contrib/instrumentation/github.com/gin-gonic/gin/otelgin"
 	"go.uber.org/fx"
 
+	"github.com/minuk-dev/opampcommander/internal/security"
 	"github.com/minuk-dev/opampcommander/pkg/apiserver/config"
+	"github.com/minuk-dev/opampcommander/pkg/apiserver/docs"
 )
 
 const (
@@ -63,4 +69,30 @@ func NewHTTPServer(
 	})
 
 	return srv
+}
+
+// NewEngine creates a new Gin engine and registers the provided controllers' routes.
+func NewEngine(
+	controllers []Controller,
+	securityService *security.Service,
+	logger *slog.Logger,
+) *gin.Engine {
+	engine := gin.New()
+	engine.Use(sloggin.New(logger))
+	engine.Use(gin.Recovery())
+	engine.Use(security.NewAuthJWTMiddleware(securityService))
+	engine.Use(otelgin.Middleware("opampcommander"))
+	// swagger
+	engine.GET("/swagger/*any", ginswagger.WrapHandler(swaggerfiles.Handler))
+
+	docs.SwaggerInfo.BasePath = "/"
+
+	for _, controller := range controllers {
+		routeInfo := controller.RoutesInfo()
+		for _, route := range routeInfo {
+			engine.Handle(route.Method, route.Path, route.HandlerFunc)
+		}
+	}
+
+	return engine
 }
